Break how it works page into sections that are controlled by tabs
