+++ src/pages/Staking/HowItWorks/StakingApplications/index.tsx
@@ -1,0 +1,8 @@
+import { FC } from "react"
+import { Box } from "@chakra-ui/react"
+
+const StakingApplications: FC = () => {
+  return <Box>Staking applications</Box>
+}
+
+export default StakingApplications
