+++ src/pages/Staking/HowItWorks/StakingOverview/AboutAddressesCard.tsx
