+++ src/pages/Staking/HowItWorks/StakingOverview/NewTStakesCard.tsx
@@ -1,6 +1,6 @@
 import { FC, ComponentProps } from "react"
 import { BodyMd, LabelSm, Card } from "@threshold-network/components"
-import StakingChecklist from "../../../components/StakingChecklist"
+import StakingChecklist from "../../../../components/StakingChecklist"
 
 export const NewTStakesCard: FC<ComponentProps<typeof Card>> = (props) => {
   return (
