+++ src/pages/Staking/HowItWorks/StakingOverview/StakingActionsCard.tsx
