+++ src/pages/Staking/HowItWorks/StakingOverview/ThresholdStakesCard.tsx
@@ -1,8 +1,8 @@
 import { FC, ComponentProps } from "react"
 import { UnorderedList, ListItem, useColorModeValue } from "@chakra-ui/react"
 import { BodyMd, LabelSm, Card } from "@threshold-network/components"
-import ViewInBlockExplorer from "../../../components/ViewInBlockExplorer"
-import { ExplorerDataType } from "../../../utils/createEtherscanLink"
+import ViewInBlockExplorer from "../../../../components/ViewInBlockExplorer"
+import { ExplorerDataType } from "../../../../utils/createEtherscanLink"
 
 export const ThresholdStakesCard: FC<
   ComponentProps<typeof Card> & { tStakingContractAddress: string }
