+++ src/pages/Staking/HowItWorks/StakingProviders/index.tsx
@@ -1,0 +1,13 @@
+import { FC } from "react"
+import { Box } from "@chakra-ui/react"
+import { ProvidersCard } from "./ProvidersCard"
+
+const StakingProviders: FC = () => {
+  return (
+    <Box>
+      <ProvidersCard />
+    </Box>
+  )
+}
+
+export default StakingProviders
